Add a PropertyDrawer for Settings<>
The user should never reference project-wide settings in a script. This way we can guide him to the project-wide settings.

TODO: The user can still reference a manual copy of a settings asset. Maybe this should be avoided too? At least for editor- and user-scoped settings it makes no sense to have multiple copies.